Move shelf files to models package
