+++ shelves/models/mongostore.go
@@ -1,4 +1,4 @@
-package main
+package models
 
 import (
 	"fmt"
