+++ shelves/models/shelfstore.go
@@ -1,4 +1,4 @@
-package main
+package models
 
 import "gopkg.in/mgo.v2/bson"
